Refactor article data handling by removing ArticleClient component; streamline Article page to directly fetch and display article data based on locale.
