+++ apps/front/app/blog/[slug]/page.tsx
@@ -1,12 +1,27 @@
+'use client'
+
+import { useLocale } from "@/contexts/LocaleContext";
 import { getArticle } from "@/data/loaders";
-import ArticleClient from "./ArticleClient";
+import ArticlePage from "@/pages/ArticlePage";
+import type { Article } from "@/types";
+import { useEffect, useState } from "react";
 
-export default async function Article({
-  params,
-}: {
-  params: { slug: string };
-}) {
-  const data = await getArticle(params.slug, "fr");
+export default function Article({ slug }: { slug: string }) {
+    const { locale } = useLocale();
+    const [data, setData] = useState<Article | null>(null);
 
-  return <ArticleClient initialData={data} slug={params.slug} />;
+    useEffect(() => {
+      const fetchData = async () => {
+        const result = await getArticle(slug, locale);
+
+        console.log("result", result);
+        setData(result);
+      };
+
+      fetchData();
+    }, [locale, slug]);
+
+    if (!data) return null;
+
+    return <ArticlePage data={data} />;
 }
